requirements.txt
# requirements.txt
torch>=1.9.0
transformers>=4.10.0
datasets>=1.11.0
sentence-transformers>=2.0.0
wandb>=0.12.0
rich>=10.0.0
nltk>=3.6.0
rouge-score>=0.0.4
scikit-learn>=0.24.0
numpy>=1.19.0
pyyaml>=5.4.0